Add sample warehouse project for event sourced library
