+++ samples/EventSourced.Sample.Warehouse.API/EventSourced.Sample.Warehouse.API.csproj
@@ -1,0 +1,20 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Controllers" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\src\EventSourced.Persistence.InMemory\EventSourced.Persistence.InMemory.csproj" />
+      <ProjectReference Include="..\EventSourced.Sample.Warehouse.Application\EventSourced.Sample.Warehouse.Application.csproj" />
+    </ItemGroup>
+
+</Project>
