+++ samples/EventSourced.Sample.Warehouse.API/Program.cs
@@ -7,7 +7,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace EventSourced.Sample.Warehouse
+namespace EventSourced.Sample.Warehouse.API
 {
     public class Program
     {
