+++ samples/EventSourced.Sample.Warehouse.API/appsettings.Development.json
