+++ samples/EventSourced.Sample.Warehouse.API/appsettings.json
