+++ samples/EventSourced.Sample.Warehouse.Application/EventSourced.Sample.Warehouse.Application.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\EventSourced.Sample.Warehouse.Domain\EventSourced.Sample.Warehouse.Domain.csproj" />
+    </ItemGroup>
+
+</Project>
