--- samples/EventSourced.Sample.Warehouse/EventSourced.Sample.Warehouse.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-    </PropertyGroup>
-
-</Project>
